LayoutTests/fast/css/fixed-background-support-enabled-expected.txt

Tests that background-attachment:fixed returns YES to @supports() when CSSFixedBackgroundSupportEnabled is enabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'backgroundAttachment' in document.documentElement.style is true
PASS 'background-attachment' in getComputedStyle(document.documentElement) is true
PASS CSS.supports('background-attachment:fixed') is true
PASS CSS.supports('background-attachment:scroll') is true
PASS CSS.supports('background-attachment:local') is true
PASS successfullyParsed is true

TEST COMPLETE
